requirementsold.txt
asgiref==3.5.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.32.0
contourpy==1.0.7
cycler==0.11.0
dj-database-url==1.2.0
Django==4.1.6
django-active-link==0.1.8
django-bootstrap-icons==0.8.2
django-cloudinary-storage==0.3.0
django-filter==22.1
django-multiselectfield==0.1.12
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-filters==0.11.1
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
Pillow==9.3.0
pillow-heif==0.7.2
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
pyzbar==0.1.9
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
urllib3==1.26.14
whitenoise==6.2.0
